requirements.txt
telethon
python-dotenv
pandas
sqlalchemy
psycopg2-binary
dbt-postgres
ultralytics
fastapi
uvicorn
dagster
dagster-webserver
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
python-dotenv
